+++ bootstrap/bukkit/src/main/java/org/geysermc/platform/bukkit/GeyserBukkitConfiguration.java
@@ -25,120 +25,29 @@
 
 package org.geysermc.platform.bukkit;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Getter;
 import org.bukkit.Bukkit;
-import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.plugin.Plugin;
 import org.geysermc.connector.FloodgateKeyLoader;
-import org.geysermc.connector.GeyserConfiguration;
+import org.geysermc.connector.configuration.GeyserJacksonConfiguration;
 
-import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.HashMap;
-import java.util.Map;
 
-public class GeyserBukkitConfiguration implements GeyserConfiguration {
+@Getter
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class GeyserBukkitConfiguration extends GeyserJacksonConfiguration {
 
-    private FileConfiguration config;
-    private File dataFolder;
-
-    private BukkitBedrockConfiguration bedrockConfig;
-    private BukkitRemoteConfiguration remoteConfig;
-    private BukkitMetricsInfo metricsInfo;
-
-    private Map<String, BukkitUserAuthenticationInfo> userAuthInfo = new HashMap<>();
+    @JsonProperty("floodgate-key-file")
+    private String floodgateKeyFile;
 
     private Path floodgateKey;
 
-    public GeyserBukkitConfiguration(File dataFolder, FileConfiguration config) {
-        this.dataFolder = dataFolder;
-        this.config = config;
-
-        bedrockConfig = new BukkitBedrockConfiguration();
-        remoteConfig = new BukkitRemoteConfiguration();
-        metricsInfo = new BukkitMetricsInfo();
-
-        if (!config.contains("userAuths"))
-            return;
-
-        for (String key : config.getConfigurationSection("userAuths").getKeys(false)) {
-            userAuthInfo.put(key, new BukkitUserAuthenticationInfo(key));
-        }
-    }
-
     public void loadFloodgate(GeyserBukkitPlugin plugin) {
         Plugin floodgate = Bukkit.getPluginManager().getPlugin("floodgate-bukkit");
-        floodgateKey = FloodgateKeyLoader.getKey(plugin.getGeyserLogger(), this, Paths.get(dataFolder.toString(), config.getString("floodgate-key-file", "public-key.pem")), floodgate, floodgate != null ? floodgate.getDataFolder().toPath() : null);
-    }
-
-    @Override
-    public IBedrockConfiguration getBedrock() {
-        return bedrockConfig;
-    }
-
-    @Override
-    public IRemoteConfiguration getRemote() {
-        return remoteConfig;
-    }
-
-    @Override
-    public Map<String, BukkitUserAuthenticationInfo> getUserAuths() {
-        return userAuthInfo;
-    }
-
-    @Override
-    public boolean isCommandSuggestions() {
-        return config.getBoolean("command-suggestions", true);
-    }
-
-    @Override
-    public boolean isPassthroughMotd() {
-        return config.getBoolean("passthrough-motd", false);
-    }
-
-    @Override
-    public boolean isPassthroughPlayerCounts() {
-        return config.getBoolean("passthrough-player-counts", false);
-    }
-
-    @Override
-    public boolean isLegacyPingPassthrough() {
-        return config.getBoolean("legacy-ping-passthrough", false);
-    }
-
-    @Override
-    public int getPingPassthroughInterval() {
-        return config.getInt("ping-passthrough-interval", 3);
-    }
-
-    @Override
-    public int getMaxPlayers() {
-        return config.getInt("max-players", 10);
-    }
-
-    @Override
-    public boolean isDebugMode() {
-        return config.getBoolean("debug-mode", false);
-    }
-
-    @Override
-    public int getGeneralThreadPool() {
-        return config.getInt("general-thread-pool", 32);
-    }
-
-    @Override
-    public boolean isAllowThirdPartyCapes() {
-        return config.getBoolean("allow-third-party-capes", true);
-    }
-
-    @Override
-    public boolean isAllowThirdPartyEars() {
-        return config.getBoolean("allow-third-party-ears", false);
-    }
-
-    @Override
-    public String getDefaultLocale() {
-        return config.getString("default-locale", "en_us");
+        floodgateKey = FloodgateKeyLoader.getKey(plugin.getGeyserLogger(), this, Paths.get(plugin.getDataFolder().toString(), plugin.getConfig().getString("floodgate-key-file", "public-key.pem")), floodgate, floodgate != null ? floodgate.getDataFolder().toPath() : null);
     }
 
     @Override
@@ -149,93 +58,5 @@ public class GeyserBukkitConfiguration implements GeyserConfiguration {
     @Override
     public boolean isCacheChunks() {
         return true; // We override this as with Bukkit, we have direct access to the server implementation
-    }
-
-    @Override
-    public boolean isAboveBedrockNetherBuilding() {
-        return config.getBoolean("above-bedrock-nether-building", false);
-    }
-
-    @Override
-    public IMetricsInfo getMetrics() {
-        return metricsInfo;
-    }
-
-    public class BukkitBedrockConfiguration implements IBedrockConfiguration {
-
-        @Override
-        public String getAddress() {
-            return config.getString("bedrock.address", "0.0.0.0");
-        }
-
-        @Override
-        public int getPort() {
-            return config.getInt("bedrock.port", 25565);
-        }
-
-        @Override
-        public String getMotd1() {
-            return config.getString("bedrock.motd1", "GeyserMC");
-        }
-
-        @Override
-        public String getMotd2() {
-            return config.getString("bedrock.motd2", "GeyserMC");
-        }
-    }
-
-    public class BukkitRemoteConfiguration implements IRemoteConfiguration {
-
-        @Override
-        public String getAddress() {
-            return config.getString("remote.address", "127.0.0.1");
-        }
-
-        @Override
-        public int getPort() {
-            return config.getInt("remote.port", 25565);
-        }
-
-        @Override
-        public String getAuthType() {
-            return config.getString("remote.auth-type", "online");
-        }
-    }
-
-    public class BukkitUserAuthenticationInfo implements IUserAuthenticationInfo {
-
-        private String key;
-
-        public BukkitUserAuthenticationInfo(String key) {
-            this.key = key;
-        }
-
-        @Override
-        public String getEmail() {
-            return config.getString("userAuths." + key + ".email");
-        }
-
-        @Override
-        public String getPassword() {
-            return config.getString("userAuths." + key + ".password");
-        }
-    }
-
-    public class BukkitMetricsInfo implements IMetricsInfo {
-
-        @Override
-        public boolean isEnabled() {
-            return config.getBoolean("metrics.enabled", true);
-        }
-
-        @Override
-        public String getUniqueId() {
-            return config.getString("metrics.uuid", "generateduuid");
-        }
-    }
-
-    @Override
-    public int getConfigVersion() {
-        return config.getInt("config-version", 0);
     }
 }
