+++ bootstrap/bungeecord/src/main/java/org/geysermc/platform/bungeecord/GeyserBungeePlugin.java
@@ -31,20 +31,19 @@ import net.md_5.bungee.config.Configuration;
 import net.md_5.bungee.config.ConfigurationProvider;
 import net.md_5.bungee.config.YamlConfiguration;
 import org.geysermc.common.PlatformType;
-import org.geysermc.connector.GeyserConfiguration;
 import org.geysermc.connector.GeyserConnector;
 import org.geysermc.connector.bootstrap.GeyserBootstrap;
 import org.geysermc.connector.command.CommandManager;
+import org.geysermc.connector.configuration.GeyserConfiguration;
 import org.geysermc.connector.ping.GeyserLegacyPingPassthrough;
 import org.geysermc.connector.ping.IGeyserPingPassthrough;
+import org.geysermc.connector.utils.FileUtils;
 import org.geysermc.platform.bungeecord.command.GeyserBungeeCommandExecutor;
 import org.geysermc.platform.bungeecord.command.GeyserBungeeCommandManager;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.InetSocketAddress;
-import java.nio.file.Files;
 import java.util.UUID;
 import java.util.logging.Level;
 
@@ -62,31 +61,17 @@ public class GeyserBungeePlugin extends Plugin implements GeyserBootstrap {
         if (!getDataFolder().exists())
             getDataFolder().mkdir();
 
-        File file = new File(getDataFolder(), "config.yml");
         Configuration configuration = null;
-
-        if (!file.exists()) {
-            try (InputStream in = getResourceAsStream("config.yml")) {
-                Files.copy(in, file.toPath());
-            } catch (IOException ex) {
-                getLogger().log(Level.SEVERE, "Failed to read/create config.yml! Make sure it's up to date and/or readable+writable!", ex);
-                return;
-            }
-        }
         try {
+            if (!getDataFolder().exists())
+                getDataFolder().mkdir();
+            File configFile = FileUtils.fileOrCopiedFromResource(new File(getDataFolder(), "config.yml"), "config.yml", (x) -> x.replaceAll("generateduuid", UUID.randomUUID().toString()));
+            this.geyserConfig = FileUtils.loadConfig(configFile, GeyserBungeeConfiguration.class);
             configuration = ConfigurationProvider.getProvider(YamlConfiguration.class).load(new File(getDataFolder(), "config.yml"));
-        } catch(IOException e) {
-            e.printStackTrace();
+        } catch (IOException ex) {
+            getLogger().log(Level.WARNING, "Failed to read/create config.yml! Make sure it's up to date and/or readable+writable!", ex);
+            ex.printStackTrace();
         }
-
-        if (configuration == null) {
-            getLogger().severe("Failed to read/create config.yml! Make sure it's up to date and/or readable+writable!");
-            return;
-        }
-
-        this.geyserConfig = new GeyserBungeeConfiguration(getDataFolder(), configuration);
-
-        boolean configHasChanged = false;
 
         if (getProxy().getConfig().getListeners().size() == 1) {
             ListenerInfo listener = getProxy().getConfig().getListeners().toArray(new ListenerInfo[0])[0];
@@ -96,33 +81,21 @@ public class GeyserBungeePlugin extends Plugin implements GeyserBootstrap {
             // Don't change the ip if its listening on all interfaces
             // By default this should be 127.0.0.1 but may need to be changed in some circumstances
             if (!javaAddr.getHostString().equals("0.0.0.0") && !javaAddr.getHostString().equals("")) {
-                configuration.set("remote.address", javaAddr.getHostString());
+                this.geyserConfig.getRemote().setAddress(javaAddr.getHostString());
             }
 
-            configuration.set("remote.port", javaAddr.getPort());
-
-            configHasChanged = true;
-        }
-
-        if (geyserConfig.getMetrics().getUniqueId().equals("generateduuid")) {
-            configuration.set("metrics.uuid", UUID.randomUUID().toString());
-
-            configHasChanged = true;
-        }
-
-        if (configHasChanged) {
-            try {
-                ConfigurationProvider.getProvider(YamlConfiguration.class).save(configuration, new File(getDataFolder(), "config.yml"));
-            } catch (IOException ex) {
-                getLogger().log(Level.SEVERE, "Failed to read/create config.yml! Make sure it's up to date and/or readable+writable!", ex);
-                return;
-            }
+            this.geyserConfig.getRemote().setPort(javaAddr.getPort());
         }
 
         this.geyserLogger = new GeyserBungeeLogger(getLogger(), geyserConfig.isDebugMode());
         GeyserConfiguration.checkGeyserConfiguration(geyserConfig, geyserLogger);
 
-        geyserConfig.loadFloodgate(this);
+        if (geyserConfig.getRemote().getAuthType().equals("floodgate") && getProxy().getPluginManager().getPlugin("floodgate-bungee") == null) {
+            geyserLogger.severe("Auth type set to Floodgate but Floodgate not found! Disabling...");
+            return;
+        }
+
+        geyserConfig.loadFloodgate(this, configuration);
 
         this.connector = GeyserConnector.start(PlatformType.BUNGEECORD, this);
