+++ connector/src/main/java/org/geysermc/connector/FloodgateKeyLoader.java
@@ -26,6 +26,8 @@
 
 package org.geysermc.connector;
 
+import org.geysermc.connector.configuration.GeyserConfiguration;
+
 import java.nio.file.Files;
 import java.nio.file.Path;
