+++ connector/src/main/java/org/geysermc/connector/ping/GeyserLegacyPingPassthrough.java
@@ -33,9 +33,7 @@ import com.github.steveice10.mc.protocol.data.status.handler.ServerInfoHandler;
 import com.github.steveice10.packetlib.Client;
 import com.github.steveice10.packetlib.tcp.TcpSessionFactory;
 import org.geysermc.common.ping.GeyserPingInfo;
-import org.geysermc.connector.GeyserConfiguration;
 import org.geysermc.connector.GeyserConnector;
-import org.geysermc.connector.GeyserLogger;
 
 import java.util.concurrent.TimeUnit;
